requirements.txt
pyparted==3.11.2
